feat: Add admin user and guest book seeder
This commit adds an admin user and a guest book seeder to the database seeder. The admin user has the name "Admin" and the email "admin@admin.com". The guest book seeder is called using the `GuestBookSeeder` class. This change is made to populate the database with initial data for testing and development purposes.