feat: Add searchable functionality to name and email columns in GuestBookResource table
This commit adds the `searchable()` method to the `TextColumn` definitions for the "name" and "email" columns in the GuestBookResource table. By making these columns searchable, users can easily find specific entries in the guest book table based on the name or email. This improves the usability and searchability of the guest book feature.